+++ init/globalinit/tradecond.go
@@ -9,7 +9,7 @@ import (
 func init() {
 	global.ForwardCond = simulationcond.AnalyzeCondition{
 		TrimHistoryCloseCount: true,
-		HistoryCloseCount:     2300,
+		HistoryCloseCount:     1900,
 		OutInRatio:            85,
 		ReverseOutInRatio:     3,
 		CloseDiff:             0,
@@ -20,15 +20,15 @@ func init() {
 		RsiLow:                0.2,
 		ReverseRsiHigh:        0.8,
 		ReverseRsiLow:         0.2,
-		TicksPeriodThreshold:  4,
-		TicksPeriodLimit:      4 * 1.3,
+		TicksPeriodThreshold:  8,
+		TicksPeriodLimit:      8 * 1.3,
 		TicksPeriodCount:      2,
-		VolumePerSecond:       30,
+		VolumePerSecond:       20,
 	}
 
 	global.ReverseCond = simulationcond.AnalyzeCondition{
 		TrimHistoryCloseCount: true,
-		HistoryCloseCount:     1200,
+		HistoryCloseCount:     800,
 		OutInRatio:            85,
 		ReverseOutInRatio:     3,
 		CloseDiff:             0,
@@ -36,12 +36,12 @@ func init() {
 		CloseChangeRatioHigh:  3,
 		OpenChangeRatio:       3,
 		RsiHigh:               0.7,
-		RsiLow:                0.2,
+		RsiLow:                0.1,
 		ReverseRsiHigh:        0.7,
-		ReverseRsiLow:         0.2,
-		TicksPeriodThreshold:  4,
-		TicksPeriodLimit:      4 * 1.3,
-		TicksPeriodCount:      2,
+		ReverseRsiLow:         0.1,
+		TicksPeriodThreshold:  8,
+		TicksPeriodLimit:      8 * 1.3,
+		TicksPeriodCount:      1,
 		VolumePerSecond:       30,
 	}
 }