+++ pkg/models/streamtick/access.go
@@ -28,3 +28,14 @@ func InsertMultiRecord(tickArr []*StreamTick, db *gorm.DB) error {
 	})
 	return err
 }
+
+// DeleteAll DeleteAll
+func DeleteAll(db *gorm.DB) error {
+	err := db.Transaction(func(tx *gorm.DB) error {
+		if err := tx.Not("id = 0").Unscoped().Delete(&StreamTick{}).Error; err != nil {
+			return err
+		}
+		return nil
+	})
+	return err
+}