+++ pkg/modules/tradebot/quota.go
@@ -14,9 +14,6 @@ import (
 // TradeQuota TradeQuota
 var TradeQuota int64 = 1000000
 
-// StatusFirstBack StatusFirstBack
-var StatusFirstBack bool
-
 const (
 	// TradeFeeRatio TradeFeeRatio
 	TradeFeeRatio float64 = 0.001425
@@ -28,85 +25,81 @@ const (
 
 // InitStartUpQuota InitStartUpQuota
 func InitStartUpQuota() {
-	for {
-		if StatusFirstBack {
-			realOrderArr, err := traderecord.GetAllorderByDayTime(global.TradeDay, database.GetAgent())
-			if err != nil {
-				panic(err)
+	if CheckIsOpenTime() {
+		realOrderArr, err := traderecord.GetAllorderByDayTime(global.TradeDay, database.GetAgent())
+		if err != nil {
+			panic(err)
+		}
+		for _, v := range realOrderArr {
+			if v.Status != 6 {
+				continue
 			}
-			for _, v := range realOrderArr {
-				if v.Status != 6 {
-					continue
-				}
-				record := traderecord.TradeRecord{
-					StockNum:  v.StockNum,
-					StockName: global.AllStockNameMap.GetName(v.StockNum),
-					Action:    v.Action,
-					Price:     v.Price,
-					Quantity:  v.Quantity,
-					Status:    v.Status,
-					OrderID:   v.OrderID,
-					TradeTime: time.Now(),
-				}
-				if v.Action == 1 {
-					switch {
-					case !FilledBuyOrderMap.CheckStockExist(v.StockNum) && !FilledBuyLaterOrderMap.CheckStockExist(v.StockNum):
-						if FilledSellFirstOrderMap.CheckStockExist(v.StockNum) {
-							FilledBuyLaterOrderMap.Set(record)
-							logger.GetLogger().Warnf("Filled Buy Later: %s", record.StockNum)
-						} else {
-							FilledBuyOrderMap.Set(record)
-							TradeQuota -= GetStockBuyCost(v.Price, v.Quantity)
-							logger.GetLogger().Warnf("Filled Buy: %s", record.StockNum)
-						}
-					case !FilledBuyLaterOrderMap.CheckStockExist(v.StockNum) && FilledBuyOrderMap.CheckStockExist(v.StockNum):
+			record := traderecord.TradeRecord{
+				StockNum:  v.StockNum,
+				StockName: global.AllStockNameMap.GetName(v.StockNum),
+				Action:    v.Action,
+				Price:     v.Price,
+				Quantity:  v.Quantity,
+				Status:    v.Status,
+				OrderID:   v.OrderID,
+				TradeTime: time.Now(),
+			}
+			if v.Action == 1 {
+				switch {
+				case !FilledBuyOrderMap.CheckStockExist(v.StockNum) && !FilledBuyLaterOrderMap.CheckStockExist(v.StockNum):
+					if FilledSellFirstOrderMap.CheckStockExist(v.StockNum) {
 						FilledBuyLaterOrderMap.Set(record)
 						logger.GetLogger().Warnf("Filled Buy Later: %s", record.StockNum)
-					case FilledBuyLaterOrderMap.CheckStockExist(v.StockNum) && !FilledBuyOrderMap.CheckStockExist(v.StockNum):
+					} else {
 						FilledBuyOrderMap.Set(record)
 						TradeQuota -= GetStockBuyCost(v.Price, v.Quantity)
 						logger.GetLogger().Warnf("Filled Buy: %s", record.StockNum)
 					}
-				} else if v.Action == 2 {
-					switch {
-					case !FilledSellOrderMap.CheckStockExist(v.StockNum) && !FilledSellFirstOrderMap.CheckStockExist(v.StockNum):
-						if FilledBuyOrderMap.CheckStockExist(v.StockNum) {
-							FilledSellOrderMap.Set(record)
-							logger.GetLogger().Warnf("Filled Sell: %s", record.StockNum)
-						} else {
-							FilledSellFirstOrderMap.Set(record)
-							TradeQuota -= GetStockBuyCost(v.Price, v.Quantity)
-							logger.GetLogger().Warnf("Filled Sell First %s", record.StockNum)
-						}
-					case !FilledSellOrderMap.CheckStockExist(v.StockNum) && FilledSellFirstOrderMap.CheckStockExist(v.StockNum):
+				case !FilledBuyLaterOrderMap.CheckStockExist(v.StockNum) && FilledBuyOrderMap.CheckStockExist(v.StockNum):
+					FilledBuyLaterOrderMap.Set(record)
+					logger.GetLogger().Warnf("Filled Buy Later: %s", record.StockNum)
+				case FilledBuyLaterOrderMap.CheckStockExist(v.StockNum) && !FilledBuyOrderMap.CheckStockExist(v.StockNum):
+					FilledBuyOrderMap.Set(record)
+					TradeQuota -= GetStockBuyCost(v.Price, v.Quantity)
+					logger.GetLogger().Warnf("Filled Buy: %s", record.StockNum)
+				}
+			} else if v.Action == 2 {
+				switch {
+				case !FilledSellOrderMap.CheckStockExist(v.StockNum) && !FilledSellFirstOrderMap.CheckStockExist(v.StockNum):
+					if FilledBuyOrderMap.CheckStockExist(v.StockNum) {
 						FilledSellOrderMap.Set(record)
 						logger.GetLogger().Warnf("Filled Sell: %s", record.StockNum)
-					case FilledSellOrderMap.CheckStockExist(v.StockNum) && !FilledSellFirstOrderMap.CheckStockExist(v.StockNum):
+					} else {
 						FilledSellFirstOrderMap.Set(record)
 						TradeQuota -= GetStockBuyCost(v.Price, v.Quantity)
 						logger.GetLogger().Warnf("Filled Sell First %s", record.StockNum)
 					}
+				case !FilledSellOrderMap.CheckStockExist(v.StockNum) && FilledSellFirstOrderMap.CheckStockExist(v.StockNum):
+					FilledSellOrderMap.Set(record)
+					logger.GetLogger().Warnf("Filled Sell: %s", record.StockNum)
+				case FilledSellOrderMap.CheckStockExist(v.StockNum) && !FilledSellFirstOrderMap.CheckStockExist(v.StockNum):
+					FilledSellFirstOrderMap.Set(record)
+					TradeQuota -= GetStockBuyCost(v.Price, v.Quantity)
+					logger.GetLogger().Warnf("Filled Sell First %s", record.StockNum)
 				}
 			}
-			logger.GetLogger().Warnf("Initial Quota: %d", TradeQuota)
-			findUnfinishedStock()
-			break
 		}
-		time.Sleep(5 * time.Second)
+		logger.GetLogger().Warnf("Initial Quota: %d", TradeQuota)
 	}
 }
 
-func findUnfinishedStock() {
+// FindUnfinishedStock FindUnfinishedStock
+func FindUnfinishedStock() {
 	filledBuyOrder := FilledBuyOrderMap.GetAllRecordMap()
 	for stockNum, record := range filledBuyOrder {
-		if !FilledSellOrderMap.CheckStockExist(stockNum) {
+		if !FilledSellOrderMap.CheckStockExist(stockNum) && !BuyOrderMap.CheckStockExist(stockNum) {
 			BuyOrderMap.Set(record)
 			logger.GetLogger().Warnf("Unfinished buy %s", record.StockNum)
 		}
 	}
 	filledSellFirstOrder := FilledSellFirstOrderMap.GetAllRecordMap()
 	for stockNum, record := range filledSellFirstOrder {
-		if !FilledBuyLaterOrderMap.CheckStockExist(stockNum) {
+		if !FilledBuyLaterOrderMap.CheckStockExist(stockNum) && !SellFirstOrderMap.CheckStockExist(stockNum) {
 			SellFirstOrderMap.Set(record)
 			logger.GetLogger().Warnf("Unfinished sell first %s", record.StockNum)
 		}