+++ pkg/modules/tradebot/reverse.go
@@ -93,7 +93,8 @@ func BuyLaterBot(ch chan *streamtick.StreamTick, cond simulationcond.AnalyzeCond
 
 // IsSellFirstPoint IsSellFirstPoint
 func IsSellFirstPoint(analyzeTick *analyzestreamtick.AnalyzeStreamTick, cond simulationcond.AnalyzeCondition) bool {
-	if analyzeTick.OpenChangeRatio > cond.OpenChangeRatio {
+	closeChangeRatio := analyzeTick.CloseChangeRatio
+	if analyzeTick.OpenChangeRatio > cond.OpenChangeRatio || closeChangeRatio > cond.CloseChangeRatioHigh {
 		return false
 	}
 	if analyzeTick.Volume < cond.VolumePerSecond*int64(analyzeTick.TotalTime) {
@@ -159,8 +160,15 @@ func CheckSellFirstOrderStatus(record traderecord.TradeRecord) {
 		}
 		if record.TradeTime.Add(tradeInWaitTime).Before(time.Now()) {
 			if err := Cancel(record.OrderID); err != nil {
-				logger.GetLogger().Error(err)
+				logger.GetLogger().WithFields(map[string]interface{}{
+					"StockNum": order.StockNum,
+					"Name":     order.StockName,
+					"Quantity": order.Quantity,
+					"Price":    order.Price,
+					"Error":    err,
+				}).Error("Cancel Fail")
 			}
+			time.Sleep(5 * time.Second)
 		}
 	}
 }
@@ -196,8 +204,15 @@ func CheckBuyLaterOrderStatus(record traderecord.TradeRecord) {
 		}
 		if record.TradeTime.Add(tradeOutWaitTime).Before(time.Now()) {
 			if err := Cancel(record.OrderID); err != nil {
-				logger.GetLogger().Error(err)
+				logger.GetLogger().WithFields(map[string]interface{}{
+					"StockNum": order.StockNum,
+					"Name":     order.StockName,
+					"Quantity": order.Quantity,
+					"Price":    order.Price,
+					"Error":    err,
+				}).Error("Cancel Fail")
 			}
+			time.Sleep(5 * time.Second)
 		}
 	}
 }